started template creation. did controller mapping manager generic, manager generic,  ssManager generic.
need to implement the handlers. basically every handler can have a class template_header field that will hold a description of that handler instructions of how to get datafiles or use thhe template etc.
also added a populate_template file that uses a load_schema function that can be overloaded to fill in the actual data. by default it's just empty